[WIP] Add a controller and an update method in PopUpFrame
The best way to handle the data exchanges between the file explorer and the frame is, to me, to add a controller. That way the file explorer and the frame can be independent (no reference between the objects) and you can provide a method in each one of the classes to monitor the exchanges between the objects.